sol color

+++ src/main/java/com/spring/jwt/entity/SalesPerson.java
@@ -27,7 +27,7 @@ public class SalesPerson {
 
 
     @Column(name = "document_id")
-    private Integer documentId;
+    private Long documentId;
 
     @Column(name = "area")
     private String area;
